requirements.txt
addict==2.4.0
aenum==3.1.16
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.5.0
anyio==3.7.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
conda==4.10.1
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262151086/work
crcmod==1.7
cryptography @ file:///tmp/build/80754af9/cryptography_1616769182610/work
cycler==0.11.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
dill==0.3.7
dnspython==2.1.0
exceptiongroup==1.3.1
fastapi==0.103.2
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
flatbuffers==25.12.19
fonttools==4.38.0
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
grpcio==1.62.3
h11==0.14.0
humanfriendly==10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata==6.7.0
ipython @ file:///tmp/build/80754af9/ipython_1617118429768/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
jmespath==0.10.0
kiwisolver==1.4.5
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==3.5.3
mdurl==0.1.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853974840/work
mkl-service==2.3.0
mmcv==2.1.0
mmdeploy==1.3.1
mmdeploy-runtime-gpu==1.3.1
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
multiprocess==0.70.15
numpy==1.21.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.19.0.56
olefile==0.46
onnx==1.14.1
onnxruntime==1.14.1
onnxsim==0.4.17
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==1.3.5
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1617386154241/work
pkginfo==1.7.0
platformdirs==4.0.0
prettytable==3.7.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.20.2
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycocotools==2.0.7
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.23.0
pydantic==2.5.3
pydantic-core==2.14.6
pygments==2.17.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
python-etcd==0.4.5
python-multipart==0.0.8
pytz==2023.4
PyYAML==6.0.1
requests==2.28.2
rich==13.8.1
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
scipy==1.7.3
shapely==2.0.7
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.3.1
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
starlette==0.27.0
sympy==1.10.1
tabulate==0.9.0
tensorrt==8.6.1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
terminaltables==3.1.10
tomli==2.0.1
torch==1.9.0
torchelastic==0.2.0
torchtext==0.10.0
torchvision==0.10.0
tqdm==4.65.2
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==4.7.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
uvicorn==0.22.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
yapf==0.43.0
zipp==3.15.0
